+++ src/ElemCo.jl
@@ -11,8 +11,10 @@ include("constants.jl")
 include("myio.jl")
 include("mnpy.jl")
 include("dump.jl")
-include("integrals.jl")
-include("msystem.jl")
+include("system/elements.jl")
+include("system/msystem.jl")
+include("system/basisset.jl")
+include("system/integrals.jl")
 
 include("ecinfos.jl")
 include("ecmethods.jl")
@@ -26,6 +28,7 @@ include("decomptools.jl")
 include("cctools.jl")
 include("dfcc.jl")
 include("cc.jl")
+include("dmrg.jl")
 include("ccdriver.jl")
 
 include("bohf.jl")
@@ -34,6 +37,10 @@ include("dfhf.jl")
 include("dfdump.jl")
 
 include("dfmcscf.jl")
+
+include("interfaces/molpro.jl")
+include("interfaces/molden.jl")
+include("interfaces/interfaces.jl")
 
 try
   using MKL
@@ -57,20 +64,25 @@ using .DFCoupledCluster
 using .FciDump
 using .DumpTools
 using .OrbTools
+using .Elements
 using .MSystem
+using .BasisSets
 using .BOHF
 using .DFHF
 using .DFMCSCF
 using .DfDump
+using .DMRG
+using .Interfaces
 
 
 export @mainname, @print_input
 export @loadfile, @savefile, @copyfile
-export @ECinit, @tryECinit, @set, @opt, @reset, @run, @method2string
-export @transform_ints, @write_ints, @dfints, @freeze_orbs, @rotate_orbs
+export @ECinit, @tryECinit, @set, @opt, @reset, @run, @var2string
+export @transform_ints, @write_ints, @dfints, @freeze_orbs, @rotate_orbs, @show_orbs
 export @dfhf, @dfuhf, @cc, @dfcc, @bohf, @bouhf, @dfmcscf
+export @import_matrix, @export_molden
 
-const __VERSION__ = "0.11.1+"
+const __VERSION__ = "0.12.0+"
 
 """
     __init__()
@@ -136,14 +148,17 @@ macro mainname(file)
 end
 
 """
-    @print_input()
+    @print_input(print_init=false)
 
   Print the input file content. 
 
   Can be used to print the input file content to the output.
 """
-macro print_input()
+macro print_input(print_init=false)
   return quote
+    if $(esc(print_init))
+      __init__()
+    end
     try
       print_info(read($(string(__source__.file)), String))
     catch
@@ -164,8 +179,10 @@ orbs = @loadfile("C_Am")
 ```
 """
 macro loadfile(filename)
+  strfilename=replace("$filename", " " => "")
   return quote
-    load($(esc(:EC)), $(esc(filename)))
+    strfilename = @var2string($(esc(filename)), $(esc(strfilename)))
+    load($(esc(:EC)), strfilename)
   end
 end
 
@@ -180,8 +197,10 @@ end
 """
 macro savefile(filename, arr, kwargs...)
   ekwa = [esc(a) for a in kwargs]
+  strfilename=replace("$filename", " " => "")
   return quote
-    save!($(esc(:EC)), $(esc(filename)), $(esc(arr)); $(ekwa...))
+    strfilename = @var2string($(esc(filename)), $(esc(strfilename)))
+    save!($(esc(:EC)), strfilename, $(esc(arr)); $(ekwa...))
   end
 end
 
@@ -195,8 +214,12 @@ end
 """
 macro copyfile(from_file, to_file, kwargs...)
   ekwa = [esc(a) for a in kwargs]
+  strfrom=replace("$from_file", " " => "")
+  strto=replace("$to_file", " " => "")
   return quote
-    copy_file!($(esc(:EC)), $(esc(from_file)), $(esc(to_file)); $(ekwa...))
+    strfrom = @var2string($(esc(from_file)), $(esc(strfrom)))
+    strto = @var2string($(esc(to_file)), $(esc(strto)))
+    copy_file!($(esc(:EC)), strfrom, strto; $(ekwa...))
   end
 end
 
@@ -212,7 +235,7 @@ end
   # Examples
 ```julia
 geometry="He 0.0 0.0 0.0"
-basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
+basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mpfit"=>"cc-pvdz-mpfit")
 @ECinit
 # output
 Occupied orbitals:[1]
@@ -339,37 +362,37 @@ macro run(method, kwargs...)
 end
 
 """
-    @method2string(method, strmethod="")
+    @var2string(var, strvar="")
 
-  Return string representation of `method`.
+  Return string representation of `var`.
 
-  If `method` is a String variable, return the value of the variable.
-  Otherwise, return the string representation of `method` (or `strmethod` if provided).
+  If `var` is a String variable, return the value of the variable.
+  Otherwise, return the string representation of `var` (or `strvar` if provided).
 
   # Examples
 ```julia
-julia> @method2string(CCSD)
+julia> @var2string(CCSD)
 "CCSD"
 julia> CCSD = "UCCSD";
-julia> @method2string(CCSD)
+julia> @var2string(CCSD)
 "UCCSD"
 ```
 """
-macro method2string(method, strmethod="")
-  if strmethod == ""
-    strmethod = replace("$method", " " => "")
+macro var2string(var, strvar="")
+  if strvar == ""
+    strvar = replace("$var", " " => "")
   end
-  varmethod = :($(esc(method)))
+  valvar = :($(esc(var)))
   return quote
     isvar = [false]
-    try @assert(typeof($(esc(method))) <: AbstractString)
+    try @assert(typeof($(esc(var))) <: AbstractString)
       isvar[1] = true
     catch
     end
     if isvar[1]
-      $varmethod
+      $valvar
     else
-      $(esc(strmethod))
+      $(esc(strvar))
     end
   end
 end
@@ -441,7 +464,7 @@ geometry="bohr
 O      0.000000000    0.000000000   -0.130186067
 H1     0.000000000    1.489124508    1.033245507
 H2     0.000000000   -1.489124508    1.033245507"
-basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
+basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mpfit"=>"cc-pvdz-mpfit")
 @dfhf
 @dfints
 @cc ccsd
@@ -453,7 +476,7 @@ macro cc(method, kwargs...)
   if kwarg_provided_in_macro(kwargs, :fcidump)
     return quote
       $(esc(:@tryECinit))
-      strmethod = @method2string($(esc(method)), $(esc(strmethod)))
+      strmethod = @var2string($(esc(method)), $(esc(strmethod)))
       ccdriver($(esc(:EC)), strmethod; $(ekwa...))
     end
   else
@@ -462,7 +485,7 @@ macro cc(method, kwargs...)
       if !fd_exists($(esc(:EC)).fd)
         $(esc(:@dfints))
       end
-      strmethod = @method2string($(esc(method)), $(esc(strmethod)))
+      strmethod = @var2string($(esc(method)), $(esc(strmethod)))
       ccdriver($(esc(:EC)), strmethod; fcidump="", $(ekwa...))
     end
   end
@@ -483,7 +506,7 @@ geometry="bohr
 O      0.000000000    0.000000000   -0.130186067
 H1     0.000000000    1.489124508    1.033245507
 H2     0.000000000   -1.489124508    1.033245507"
-basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
+basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mpfit"=>"cc-pvdz-mpfit")
 @dfhf
 @dfcc svd-dcsd
 ```
@@ -492,7 +515,7 @@ macro dfcc(method="svd-dcsd")
   strmethod=replace("$method", " " => "")
   return quote
     $(esc(:@tryECinit))
-    strmethod = @method2string($(esc(method)), $(esc(strmethod)))
+    strmethod = @var2string($(esc(method)), $(esc(strmethod)))
     dfccdriver($(esc(:EC)), strmethod)
   end
 end
@@ -559,7 +582,7 @@ macro transform_ints(type="")
       error("No FCIDump found.")
     end
     CMOr = load($(esc(:EC)), $(esc(:EC)).options.wf.orb)
-    strtype = @method2string($(esc(type)), $(esc(strtype)))
+    strtype = @var2string($(esc(type)), $(esc(strtype)))
     if strtype ∈ ["bo", "BO", "bi-orthogonal", "Bi-orthogonal", "biorth", "biorthogonal", "Biorthogonal"]
       CMOl = load($(esc(:EC)), $(esc(:EC)).options.wf.orb*$(esc(:EC)).options.wf.left)
     elseif strtype == ""
@@ -633,6 +656,54 @@ macro rotate_orbs(orb1, orb2, angle, kwargs...)
   return quote
     $(esc(:@tryECinit))
     rotate_orbs($(esc(:EC)), $(esc(orb1)), $(esc(orb2)), $(esc(angle)); $(ekwa...))
+  end
+end
+
+"""
+    @show_orbs(range=nothing)
+
+  Show orbitals in the integrals according to an array or range 
+  `range`.
+
+  # Examples
+```julia
+@dfhf
+@show_orbs 1:5
+```
+"""
+macro show_orbs(range=nothing)
+  return quote
+    $(esc(:@tryECinit))
+    show_orbitals($(esc(:EC)), $(esc(range)))
+  end
+end
+
+"""
+    @import_matrix(filename)
+
+  Import matrix from file `file`.
+
+  The type of the matrix is determined automatically.
+"""
+macro import_matrix(filename)
+  strfilename=replace("$filename", " " => "")
+  return quote
+    $(esc(:@tryECinit))
+    strfilename = @var2string($(esc(filename)), $(esc(strfilename)))
+    import_matrix($(esc(:EC)), strfilename)
+  end
+end
+
+"""
+    @export_molden(filename)
+
+  Export current orbitals to Molden file `filename`.
+"""
+macro export_molden(filename)
+  strfilename=replace("$filename", " " => "")
+  return quote
+    strfilename = @var2string($(esc(filename)), $(esc(strfilename)))
+    export_molden_orbitals($(esc(:EC)), strfilename)
   end
 end