+++ src/dfhf.jl
@@ -2,7 +2,7 @@ module DFHF
 using LinearAlgebra, TensorOperations, Printf
 using ..ElemCo.Utils
 using ..ElemCo.ECInfos
-using ..ElemCo.ECInts
+using ..ElemCo.Integrals
 using ..ElemCo.MSystem
 using ..ElemCo.OrbTools
 using ..ElemCo.DFTools
@@ -33,8 +33,8 @@ function dfhf(EC::ECInfo)
   guess = EC.options.scf.guess
   Enuc = generate_AO_DF_integrals(EC, "jkfit"; save3idx=!direct)
   if direct
-    bao = generate_basis(EC.system, "ao")
-    bfit = generate_basis(EC.system, "jkfit")
+    bao = generate_basis(EC, "ao")
+    bfit = generate_basis(EC, "jkfit")
   end
   t1 = print_time(EC, t1, "generate AO-DF integrals", 2)
   cMO = guess_orb(EC,guess)
@@ -80,6 +80,7 @@ function dfhf(EC::ECInfo)
     t1 = print_time(EC, t1, "diagonalize Fock matrix", 2)
     # display(ϵ)
   end
+  normalize_phase!(cMO)
   println("DF-HF energy: ", EHF)
   draw_endline()
   delete_temporary_files!(EC)
@@ -108,14 +109,14 @@ function dfuhf(EC::ECInfo)
   guess = EC.options.scf.guess
   Enuc = generate_AO_DF_integrals(EC, "jkfit"; save3idx=!direct)
   if direct
-    bao = generate_basis(EC.system, "ao")
-    bfit = generate_basis(EC.system, "jkfit")
+    bao = generate_basis(EC, "ao")
+    bfit = generate_basis(EC, "jkfit")
   end
   t1 = print_time(EC, t1, "generate AO-DF integrals", 2)
   cMO = guess_orb(EC,guess)
   t1 = print_time(EC, t1, "guess orbitals", 2)
   if !is_unrestricted_MO(cMO)
-    cMO = Any[cMO, cMO]
+    cMO = Array{Float64}[cMO, cMO]
   end
   ϵ = [zeros(norb), zeros(norb)] 
   hsmall = load(EC, "h_AA")
@@ -133,8 +134,8 @@ function dfuhf(EC::ECInfo)
       fock = gen_dffock(EC,cMO)
     end
     t1 = print_time(EC, t1, "generate DF-Fock matrix", 2)
-    efhsmall = Any[0.0, 0.0]
-    Δfock = Any[zeros(norb,norb), zeros(norb,norb)]
+    efhsmall = Float64[0.0, 0.0]
+    Δfock = Array{Float64}[zeros(norb,norb), zeros(norb,norb)]
     var = 0.0
     for (ispin, sp) = enumerate(['o', 'O'])
       den = gen_density_matrix(EC, cMO[ispin], cMO[ispin], SP[sp])
@@ -162,6 +163,9 @@ function dfuhf(EC::ECInfo)
     end
     t1 = print_time(EC, t1, "diagonalize Fock matrix", 2)
     # display(ϵ)
+  end
+  for ispin = 1:2
+    normalize_phase!(cMO[ispin])
   end
   println("DF-UHF energy: ", EHF)
   draw_endline()