+++ src/options.jl
@@ -36,6 +36,10 @@ Base.@kwdef mutable struct WfOptions
   occb::String = "-"
   """`⟨false⟩` ignore various errors in sanity checks. """
   ignore_error::Bool = false
+  """`⟨5⟩` number of largest orbitals to print. """
+  print_nlargest::Int = 5
+  """`⟨0.1⟩` threshold for orbital coefficients to print. """
+  print_thr::Float64 = 0.1
 end
 
 
@@ -193,6 +197,22 @@ Base.@kwdef mutable struct CcOptions
   dcsd_ofac::Float64 = 0.15
 end
 
+""" 
+  Options for DMRG calculation.
+
+  $(TYPEDFIELDS)
+"""
+Base.@kwdef mutable struct DmrgOptions
+  """`⟨10⟩` number of sweeps. """
+  nsweeps::Int = 10
+  """`⟨[100, 200]⟩` maximum size for the bond dimension. """
+  maxdim::Vector{Int} = [100, 200]
+  """`⟨1e-6⟩` cutoff for the singular value decomposition. """
+  cutoff::Float64 = 1e-6
+  """`⟨[1e-6, 1e-7, 1e-8, 0.0]⟩` strength of the noise term used to aid convergence. """
+  noise::Vector{Float64} = [1e-6, 1e-7, 1e-8, 0.0]
+end
+
 """
   Options for integral calculation.
 
@@ -203,6 +223,8 @@ Base.@kwdef mutable struct IntOptions
   df::Bool = true
   """`⟨""⟩` store integrals in FCIDump format. """
   fcidump::String = ""
+  """`⟨false⟩` use Cartesian subshells instead of Spherical. """
+  cartesian::Bool = false
 end
 
 """ 
@@ -248,6 +270,8 @@ Base.@kwdef mutable struct Options
   int::IntOptions = IntOptions()
   """ Coupled-Cluster options ([`CcOptions`](@ref)). """
   cc::CcOptions = CcOptions()
+  """ DMRG options ([`DmrgOptions`](@ref)). """
+  dmrg::DmrgOptions = DmrgOptions()
   """ Cholesky options ([`CholeskyOptions`](@ref)). """
   cholesky::CholeskyOptions = CholeskyOptions()
   """ DIIS options ([`DiisOptions`](@ref)). """