+++ test/df_uhf.jl
@@ -15,7 +15,7 @@ geometry="bohr
 
 basis = Dict("ao"=>"cc-pVDZ",
              "jkfit"=>"cc-pvtz-jkfit",
-             "mp2fit"=>"cc-pvdz-rifit")
+             "mpfit"=>"cc-pvdz-rifit")
 let
   @opt wf ms2=2
   EUHF = @dfuhf