+++ test/h2o_udcsd_prop.jl
@@ -14,7 +14,7 @@ geometry="bohr
 
 basis = Dict("ao"=>"cc-pVDZ",
              "jkfit"=>"cc-pvdz-rifit",
-             "mp2fit"=>"cc-pvdz-rifit")
+             "mpfit"=>"cc-pvdz-rifit")
 let
   @opt wf ms2=2
   @dfuhf